requirements.txt
aiogram>=3.15,<4.0
python-dotenv>=1.0,<2.0
